coming soon

+++ minitalk.h
@@ -6,7 +6,7 @@
 /*   By: fel-abbo <fel-abbo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/18 03:45:21 by fel-abbo          #+#    #+#             */
-/*   Updated: 2024/04/25 07:33:36 by fel-abbo         ###   ########.fr       */
+/*   Updated: 2024/04/26 10:01:46 by fel-abbo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,6 +27,6 @@ int			ft_isdigit(int c);
 int			ft_isprint(int c);
 void 		ft_putchar(char c);
 void 		ft_putstr(char *str);
-void 		ft_putnbr(int nb);
+void 		ft_putnbr(long int nb);
 
 #endif
